src/requirements/base.txt
Django==5.1.5
djangorestframework==3.16.0
python-dotenv==1.1.0
django-cors-headers==4.7.0
django-filter==25.1
django-extensions==4.1
requests==2.32.3
httptools==0.6.4
whitenoise==6.9.0
PyJWT==2.6.0
Pillow==11.2.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-import-export==4.3.7
drf-yasg==1.21.10
redis==6.1.0
django-redis==5.4.0
boto3==1.38.21
django-storages==1.14.6
shortuuid==1.0.13
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
websockets==12.0
sentry-sdk==2.8.0
celery==5.5.3
django-celery-beat==2.7.0
google-generativeai>=0.8.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
openai>=1.12.0
beautifulsoup4==4.12.2
lxml==5.1.0
rank-bm25==0.2.2
stripe>=13.0.0
pgvector==0.3.6
tiktoken==0.8.0
django-model-utils==4.3.1
prometheus-client==0.21.0
celery-prometheus-exporter==1.7.0
psutil==6.1.1
